# frontend/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    BlockStraight:
      allOf:
      - $ref: "#/components/schemas/Straight"
      description: type for a track part
      type: object
      properties:
        blockLength:
          type: integer
          format: int32
        leftTrackBlock:
          $ref: "#/components/schemas/TrackBlock"
        middleTrackBlock:
          $ref: "#/components/schemas/TrackBlock"
        rightTrackBlock:
          $ref: "#/components/schemas/TrackBlock"
        endGridPosition:
          $ref: "#/components/schemas/GridPosition"
        gridLength:
          type: integer
          format: int32
        allTrackBlocks:
          type: array
          items:
            $ref: "#/components/schemas/TrackBlock"
...
    BusDataConfiguration:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bus:
          type: integer
          format: int32
        address:
          type: integer
          format: int32
        bit:
          type: integer
          format: int32
        bitState:
          type: boolean
...
    Construction:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        inAutomaticMode:
          type: boolean
...
    ConstructionDto:
      type: object
      properties:
        name:
          type: string
...
    DIRECTION:
      type: string
      enum:
      - HORIZONTAL
      - VERTICAL
...
    DRIVING_DIRECTION1:
      type: string
      enum:
      - FORWARD
      - BACKWARD
    DRIVING_LEVEL_ADJUST_TYPE:
      type: string
      enum:
      - NONE
      - ENTER
      - EXIT
...
    GridPosition:
      type: object
      properties:
        id:
          type: integer
          format: int64
        x:
          type: integer
          format: int32
        "y":
          type: integer
          format: int32
    LocalDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50
...
    Route:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        start:
          $ref: "#/components/schemas/BlockStraight"
        end:
          $ref: "#/components/schemas/TrackBlock"
        oneway:
          type: boolean
        waypoints:
          type: array
          items:
            $ref: "#/components/schemas/GridPosition"
        track:
          $ref: "#/components/schemas/Track"
        allTrackBlocksToDrive:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/TrackBlock"
...
    RouteSequence:
      type: object
      properties:
        id:
          type: integer
          format: int64
        position:
          type: integer
          format: int32
        route:
          $ref: "#/components/schemas/Route"
        endDelayInSeconds:
          type: integer
          format: int32
...
    STATE:
      type: string
      enum:
      - SUCCESS
      - FAILED
...
    Scenario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        cron:
          type: string
        train:
          $ref: "#/components/schemas/Train"
        trainDrivingDirection:
          $ref: "#/components/schemas/DRIVING_DIRECTION1"
        startDrivingLevel:
          type: integer
          format: int32
        routeSequences:
          type: array
          items:
            $ref: "#/components/schemas/RouteSequence"
        stationPlatformStartId:
          type: integer
          format: int64
        stationPlatformEndId:
          type: integer
          format: int64
...
    ScenarioStatistic:
      type: object
      properties:
        scenarioId:
          type: integer
          format: int64
        lastRun:
          $ref: "#/components/schemas/LocalDateTime"
        total:
          type: integer
          format: int32
        failed:
          type: integer
          format: int32
        successful:
          type: integer
          format: int32
        averageRunTimeInMillis:
          type: number
          format: double
        runs:
          type: array
          items:
            $ref: "#/components/schemas/ScenarioStatisticRun"
    ScenarioStatisticRun:
      type: object
      properties:
        start:
          $ref: "#/components/schemas/LocalDateTime"
        end:
          $ref: "#/components/schemas/LocalDateTime"
        averageRunTimeInMillis:
          type: number
          format: double
        state:
          $ref: "#/components/schemas/STATE"
...
    Straight:
      description: type for a track part
      type: object
      properties:
        id:
          type: integer
          format: int64
        gridPosition:
          $ref: "#/components/schemas/GridPosition"
        trackPartType:
          type: string
        direction:
          $ref: "#/components/schemas/DIRECTION"
      required:
      - trackPartType
...
    Track:
      type: object
      properties:
        trackBlocks:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/TrackBlock"
        trackFunctions:
          type: array
          items:
            $ref: "#/components/schemas/BusDataConfiguration"
        gridPositions:
          type: array
          items:
            $ref: "#/components/schemas/GridPosition"
        length:
          type: integer
          format: int32
    TrackBlock:
      type: object
      properties:
        id:
          type: integer
          format: int64
        blockFunction:
          $ref: "#/components/schemas/BusDataConfiguration"
        name:
          type: string
        forwardTargetDrivingLevel:
          type: integer
          format: int32
        backwardTargetDrivingLevel:
          type: integer
          format: int32
        drivingLevelAdjustType:
          $ref: "#/components/schemas/DRIVING_LEVEL_ADJUST_TYPE"
        feedback:
          type: boolean
        displayValue:
          type: string
...
    Train:
      type: object
      properties:
        id:
          type: integer
          format: int64
        address:
          type: integer
          format: int32
        name:
          type: string
        functions:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/TrainFunction"
        addressByte:
          type: integer
          format: int8
...
    TrainFunction:
      type: object
      properties:
        id:
          type: integer
          format: int64
        alias:
          type: string
        configuration:
          $ref: "#/components/schemas/BusDataConfiguration"
...
paths:
...
  /api/config/{key}:
    put:
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Update
      tags:
      - Config Resource
    get:
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
      summary: Load Value
      tags:
      - Config Resource
  /api/constructions:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Construction"
      summary: List All
      tags:
      - Construction Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConstructionDto"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Construction"
          description: Created
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Construction Resource
  /api/constructions/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConstructionDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Construction"
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Construction Resource
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Construction"
      summary: Get By Id
      tags:
      - Construction Resource
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Delete
      tags:
      - Construction Resource
...
  /api/routes:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
      summary: List All
      tags:
      - Route Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Route"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
          description: Created
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Route Resource
  /api/routes/build-track:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Route"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad Request
      summary: Build Track
      tags:
      - Route Resource
  /api/routes/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Route"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Route Resource
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
      summary: Get By Id
      tags:
      - Route Resource
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Delete
      tags:
      - Route Resource
  /api/scenarios:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
      summary: List All
      tags:
      - Scenario Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Scenario"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
          description: Created
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Scenario Resource
...
  /api/scenarios/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Scenario"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Scenario Resource
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
      summary: Get By Id
      tags:
      - Scenario Resource
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Delete
      tags:
      - Scenario Resource
  /api/scenarios/{id}/cancel:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Cancel
      tags:
      - Scenario Resource
  /api/scenarios/{id}/schedule:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Schedule
      tags:
      - Scenario Resource
  /api/scenarios/{id}/start:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Start
      tags:
      - Scenario Resource
  /api/scenarios/{id}/statistic:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenarioStatistic"
      summary: Scenario Statistic
      tags:
      - Scenario Resource
  /api/scenarios/{id}/unschedule:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Unschedule
      tags:
      - Scenario Resource
...
info:
  title: controlcenter API
  version: 0.3-SNAPSHOT